Massive module restructure DBManager removed, replaced with DB PostgresDB removed, replaced with PersonModel, BoardModel, TaskModel
