+++ example_go
@@ -1,0 +1,49 @@
+// package test
+
+// import (
+// 	"github.com/jackc/pgx/v5/pgxpool"
+// )
+
+// type Person struct {
+// 	ID   uint32
+// 	Name string
+// }
+
+// type PersonModel struct {
+// 	DB DBConn
+// }
+
+// func (pm *PersonModel) GetByID() (Person, error) {
+// 	person, err := pm.DBConn.Query("...")
+// 	if err != nil {
+// 		return Person{}, nil
+// 	}
+// 	return person, nil
+// }
+
+// type Board struct {
+// 	ID   uint32
+// 	Name string
+// }
+
+// type BoardModel struct {
+// 	DB *DBConn
+// }
+
+// func (bm *BoardModel) GetByID() (Person, error) {
+// 	board, err := bm.DBConn.Query("...")
+// 	if err != nil {
+// 		return Board{}, nil
+// 	}
+// 	return board, nil
+// }
+
+// func main() {
+// 	dbPool, err := pgxpool.New(context.Background(), dbURL)
+// 	if err != nil {
+// 		fmt.Fatal(err)
+// 	}
+// 	dbconn := DBConn(dbPool)
+// 	db := DB{Person: PersonModel{dbconn}, BoardModel{dbconn}}
+// 	db.Person.GetByID
+// }
