+++ models/subtask.go
@@ -1,0 +1,7 @@
+package models
+
+type Subtask struct {
+	Name         string
+	ID           uint32
+	ParentTaskID uint32
+}
