+++ models/tag.go
@@ -1,0 +1,8 @@
+package models
+
+type Tag struct {
+	Name        string
+	Description string
+	ID          uint32
+	BoardID     uint32
+}
